app(feature): reorganize file structure and naming
