+++ .nvmrc
@@ -1,1 +1,1 @@
-14
+18.12.1
