+++ package.json
@@ -1,6 +1,6 @@
 {
-  "name": "contentful-starter-gatsby-blog",
-  "description": "Contentful Gatsby Starter Blog",
+  "name": "misa-website-2023",
+  "description": "MISA Website 2023",
   "private": true,
   "version": "0.0.1",
   "author": "Contentful <opensource@contentful.com>",
