+++ src/components/Layout/Navbar.js
@@ -26,7 +26,7 @@ const Navigation = () => {
         <Link to="/">
           <StaticImage
             className="mx-6 my-6 ml-8 w-[143px] h-[25px]"
-            src="../../static/images/navbarlogo.png"
+            src="../../../static/images/navbarlogo.png"
             quality={100}
           />
         </Link>
