+++ src/components/Layout/index.js
@@ -1,10 +1,10 @@
 import React from 'react'
 
-import './variables.css'
-import './global.css'
-import Seo from './seo'
-import Navigation from './navigation'
-import Footer from './footer'
+import '../../../static/styles/variables.css'
+import '../../../static/styles/global.css'
+import Seo from './SEO'
+import Navigation from './Navbar'
+import Footer from './Footer'
 class Template extends React.Component {
   render() {
     const { children } = this.props
