+++ src/pages/about-us.js
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import Layout from '../components/layout'
+import Layout from '../components/Layout/index'
 import { StaticImage } from 'gatsby-plugin-image'
 
 import Button from '../components/Elements/button'
