+++ src/pages/partnerships/partners.js
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import Layout from '../../components/layout'
+import Layout from '../../components/Layout/index'
 
 const Partners = () => {
   return (
