+++ src/templates/specific-event.js
@@ -3,7 +3,7 @@ import BackgroundImage from 'gatsby-background-image'
 import { GatsbyImage } from 'gatsby-plugin-image'
 import { convertToBgImage } from 'gbimage-bridge'
 
-import Layout from '../components/layout'
+import Layout from '../components/Layout/index'
 
 import { Swiper, SwiperSlide } from 'swiper/react'
 import { Autoplay } from 'swiper'
